transformed MeasurementCard into statelessWidget and wrapped it with a selector from BmiProvider in HomeScreen

+++ lib/src/providers/bmi_provider.dart
@@ -1,17 +1,18 @@
 import 'package:bmi_calculator/src/widgets/gender_card.dart';
 import 'package:flutter/material.dart';
 
+
 class BmiProvider extends ChangeNotifier {
   // Set private BMI variables
   Gender _gender = Gender.none;
   double _height = 165; //cm
-  double _weight = 130; //lb
+  int _weight = 130; //lb
   int _age = 30; //years
 
   // Get BMI parameters from private variables
   Gender get gender => _gender;
   double get height => _height;
-  double get weight => _weight;
+  int get weight => _weight;
   int get age => _age;
 
   /// Update user gender
@@ -27,7 +28,7 @@ class BmiProvider extends ChangeNotifier {
   }
 
   /// Update user [weight]
-  void updateWeight(double newWeight) {
+  void updateWeight(int newWeight) {
     _weight = newWeight;
     notifyListeners();
   }
